Move `FrameImpl` into it's own module
