+++ packages/libs/error-stack/src/frame/mod.rs
@@ -1,162 +1,13 @@
+mod frame_impl;
 mod kind;
 
 use alloc::boxed::Box;
 #[cfg(nightly)]
 use core::any::{self, Demand, Provider};
-use core::{any::Any, fmt, panic::Location};
-use std::any::TypeId;
+use core::{any::TypeId, fmt, panic::Location};
 
+use self::frame_impl::FrameImpl;
 pub use self::kind::{AttachmentKind, FrameKind};
-use crate::Context;
-
-/// Internal representation of a [`Frame`].
-///
-/// # Safety
-///
-/// - It must be allowed to cast from `*dyn FrameImpl` to `*T` if `type_id() == TypeId::of::<T>()`.
-///   This is the case if [`type_id`] returns `TypeId::of::<T>()` and
-///     - `T` is `Self`,
-///     - `T` is the inner struct on `#[repr(transparent]`, or
-///     - `T` is the first struct on `#[repr(C)]`.
-///
-/// [`type_id`]: Self::type_id
-unsafe trait FrameImpl: Any {
-    fn kind(&self) -> FrameKind<'_>;
-
-    /// Returns the [`TypeId`] of this `Frame`.
-    ///
-    /// It's guaranteed, that `*dyn FrameImpl` can be cast to a pointer to the type returned by
-    /// `type_id`.  
-    fn type_id(&self) -> TypeId;
-
-    /// Returns the location where this `Frame` was created.
-    fn location(&self) -> &'static Location<'static>;
-
-    /// Returns a shared reference to the source of this `Frame`.
-    ///
-    /// This corresponds to the `Frame` below this one in a [`Report`].
-    ///
-    /// [`Report`]: crate::Report
-    fn source(&self) -> Option<&Frame>;
-
-    /// Returns a mutable reference to the source of this `Frame`.
-    ///
-    /// This corresponds to the `Frame` below this one in a [`Report`].
-    ///
-    /// [`Report`]: crate::Report
-    fn source_mut(&mut self) -> Option<&mut Frame>;
-
-    /// Provide values which can then be requested.
-    #[cfg(nightly)]
-    fn provide<'a>(&'a self, demand: &mut Demand<'a>);
-}
-
-#[repr(C)]
-struct ContextFrame<C> {
-    context: C,
-    location: &'static Location<'static>,
-    source: Option<Box<Frame>>,
-}
-
-// SAFETY: `type_id` returns `C` and `C` is the first field in `#[repr(C)]`
-unsafe impl<C: Context> FrameImpl for ContextFrame<C> {
-    fn kind(&self) -> FrameKind<'_> {
-        FrameKind::Context(&self.context)
-    }
-
-    fn type_id(&self) -> TypeId {
-        TypeId::of::<C>()
-    }
-
-    fn location(&self) -> &'static Location<'static> {
-        self.location
-    }
-
-    fn source(&self) -> Option<&Frame> {
-        self.source.as_ref().map(Box::as_ref)
-    }
-
-    fn source_mut(&mut self) -> Option<&mut Frame> {
-        self.source.as_mut().map(Box::as_mut)
-    }
-
-    #[cfg(nightly)]
-    fn provide<'a>(&'a self, demand: &mut Demand<'a>) {
-        Context::provide(&self.context, demand);
-    }
-}
-
-#[repr(C)]
-struct AttachmentFrame<A> {
-    attachment: A,
-    location: &'static Location<'static>,
-    source: Option<Box<Frame>>,
-}
-
-// SAFETY: `type_id` returns `A` and `A` is the first field in `#[repr(C)]`
-unsafe impl<A: 'static + Send + Sync> FrameImpl for AttachmentFrame<A> {
-    fn kind(&self) -> FrameKind<'_> {
-        FrameKind::Attachment(AttachmentKind::Opaque(&self.attachment))
-    }
-
-    fn type_id(&self) -> TypeId {
-        TypeId::of::<A>()
-    }
-
-    fn location(&self) -> &'static Location<'static> {
-        self.location
-    }
-
-    fn source(&self) -> Option<&Frame> {
-        self.source.as_ref().map(Box::as_ref)
-    }
-
-    fn source_mut(&mut self) -> Option<&mut Frame> {
-        self.source.as_mut().map(Box::as_mut)
-    }
-
-    #[cfg(nightly)]
-    fn provide<'a>(&'a self, demand: &mut Demand<'a>) {
-        demand.provide_ref(&self.attachment);
-    }
-}
-
-#[repr(C)]
-struct PrintableAttachmentFrame<A> {
-    attachment: A,
-    location: &'static Location<'static>,
-    source: Option<Box<Frame>>,
-}
-
-// SAFETY: `type_id` returns `A` and `A` is the first field in `#[repr(C)]`
-unsafe impl<A: 'static + fmt::Debug + fmt::Display + Send + Sync> FrameImpl
-    for PrintableAttachmentFrame<A>
-{
-    fn kind(&self) -> FrameKind<'_> {
-        FrameKind::Attachment(AttachmentKind::Printable(&self.attachment))
-    }
-
-    fn type_id(&self) -> TypeId {
-        TypeId::of::<A>()
-    }
-
-    fn location(&self) -> &'static Location<'static> {
-        self.location
-    }
-
-    fn source(&self) -> Option<&Frame> {
-        self.source.as_ref().map(Box::as_ref)
-    }
-
-    fn source_mut(&mut self) -> Option<&mut Frame> {
-        self.source.as_mut().map(Box::as_mut)
-    }
-
-    #[cfg(nightly)]
-    fn provide<'a>(&'a self, demand: &mut Demand<'a>) {
-        demand.provide_ref(&self.attachment);
-    }
-}
 
 /// A single context or attachment inside of a [`Report`].
 ///
@@ -174,63 +25,6 @@ pub struct Frame {
 }
 
 impl Frame {
-    /// Crates a frame from a [`Context`].
-    pub(crate) fn from_context<C>(
-        context: C,
-        location: &'static Location<'static>,
-        source: Option<Box<Self>>,
-    ) -> Self
-    where
-        C: Context,
-    {
-        Self {
-            frame: Box::new(ContextFrame {
-                context,
-                location,
-                source,
-            }),
-        }
-    }
-
-    /// Crates a frame from an attachment.
-    pub(crate) fn from_attachment<A>(
-        attachment: A,
-        location: &'static Location<'static>,
-        source: Option<Box<Self>>,
-    ) -> Self
-    where
-        A: Send + Sync + 'static,
-    {
-        Self {
-            frame: Box::new(AttachmentFrame {
-                attachment,
-                location,
-                source,
-            }),
-        }
-    }
-
-    /// Crates a frame from an attachment which implements [`Debug`] and [`Display`].
-    ///
-    /// [`Debug`]: core::fmt::Debug
-    /// [`Display`]: core::fmt::Display
-    pub(crate) fn from_printable_attachment<A>(
-        attachment: A,
-        location: &'static Location<'static>,
-        source: Option<Box<Self>>,
-    ) -> Self
-    where
-        A: fmt::Display + fmt::Debug + Send + Sync + 'static,
-    {
-        Self {
-            frame: Box::new(PrintableAttachmentFrame {
-                attachment,
-                location,
-                source,
-            }),
-        }
-    }
-
     /// Returns the location where this `Frame` was created.
     #[must_use]
     pub fn location(&self) -> &'static Location<'static> {
